requirements.txt
asgiref==3.8.1
boto3==1.37.24
botocore==1.37.24
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.0
cryptography==44.0.2
defusedxml==0.8.0rc2
Django==4.2.15
django-autoslug==1.9.8
django-environ==0.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environ==1.0
idna==3.10
iniconfig==2.1.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
Pillow==8.3.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.1.0
pytest==8.3.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.11.4
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.5.3
toml==0.10.2
uritemplate==4.1.1
urllib3==2.3.0
